+++ src/util.ts
@@ -1,0 +1,4 @@
+
+export function wrap(val: number, max: number) {
+  return ((val % max) + max) % max;
+}
